Split logic into files
